Update repository.md
Relax requirement for the code coverage badge. [There exist valid cases](https://github.com/signalfx/gdi-specification/discussions/34#discussion-3309605) where it is not possible to get an accurate measure of this value and reporting it would be misleading. In those cases we should not require the misleading information to be included.