Clarify RUM beaconEndpoint value enforcement (#146)
Some systems, Android, cannot enforce this configuration value at the
library level. It needs to be enforced at the application level.
Therefore, this part of the specification is unimplementable by the
Android RUM library. This adds an exception to this requirement to allow
the Android RUM library to be compliant.